Simplified CallKit management by removing redundent services.
